requirements.txt
langchain==0.1.8
mypy-extensions==1.0.0
numexpr==2.8.4
python-environ==0.4.54
pyodbc==4.0.39
python-dotenv==1.0.0
openai==1.10.0
huggingface-hub==0.17.3
langchain_experimental==0.0.20
flask==3.0.0
flask_cors==4.0.0
pyttsx3==2.90
wfastcgi==3.0.0
redis==5.0.1
google-generativeai==0.3.2
openpyxl==3.1.2
langchain-google-genai==0.0.9
langchain-openai==0.0.8
python-docx==0.8.11
pypdf==4.0.2
unstructured==0.10.2
networkx==3.1
langsmith==0.1.2
langchain-google-genai==0.0.9
langchain-groq
gtts
ptvsd
pytz
pandas